+++ packages/sanes-browser-extension/src/routes/share-identity/index.stories.tsx
@@ -7,7 +7,7 @@ import React from "react";
 
 import { GetIdentitiesResponseData } from "../../extension/background/model/requestsHandler/requestQueueManager";
 import { CHROME_EXTENSION_ROOT } from "../../utils/storybook";
-import { WALLET_STATUS_PAGE } from "../account/index.stories";
+import { WALLET_STATUS_PAGE } from "../wallet/index.stories";
 import RejectRequest from "./components/RejectRequest";
 import ShowRequest from "./components/ShowRequest";
