+++ packages/sanes-browser-extension/src/routes/wallet/components/Empty.tsx
