+++ packages/sanes-browser-extension/src/routes/wallet/components/Tx/MsgRegisterUsernameTx.tsx
