+++ packages/sanes-browser-extension/src/routes/wallet/components/Tx/MsgSendTransaction.tsx
