+++ packages/sanes-browser-extension/src/routes/wallet/index.e2e.spec.ts
