+++ packages/sanes-browser-extension/src/routes/wallet/index.tsx
